feat: implement admin panel
* feat: moved the admin panel from branch 16, also fixed some bugs

* feat: returned the settings back csrf in Spring Security

* feat: added button to profile to go control panel for admin and configured security config to restrict access to endpoint /users/**

* refactor: fix bug with csrf

---------

+++ src/main/java/kg/edu/mathbilim/service/impl/reference/UserTypeServiceImpl.java
@@ -22,7 +22,7 @@ public class UserTypeServiceImpl implements UserTypeService {
     }
 
     @Override
-    public UserType findById(Long id) {
+    public UserType findById(Integer id) {
         return userTypeRepository.findById(id)
                 .orElseThrow(() -> new UserTypeNotFoundException("UserType not found with id: " + id));
     }
